login system
this version of the software includes a working basic login system, which can still be improved in terms of user input validation, for allowing secure passwords.  some security features are already put on it such as [HideInInspector] which would allow only particular data to be requested and sent to the PlayFab servers, this would allow for quicker and secure communication between the client and the server.